core/pond/requirements.txt
pandas>=2.0.0
pyarrow>=14.0.1
numpy>=1.24.0
pytest>=7.4.0
scikit-learn>=1.3.0
pymysql>=1.1.0
boto3>=1.33.0
python-dotenv>=1.0.0
fsspec
s3fs
